⚡️ [WIP] Improve memory usage
This allows us to make it past 500 rounds, but it's pretty slow:

- 600 rounds takes 10 seconds
- 700 takes 80 seconds

At this point, it becomes clear that

> you'll need to find another way to keep your worry levels manageable

which, like, I know, jeez.

I have a feeling that I need to use all of the monkey's divisors but I
don't really understand how, so I spend an evening fumbling around,
trying every optimization I can think of for no real benefit, and trying
to scale the worry level down for storage and back up for inspection
(with all sorts of different factors and scales), which of course
doesn't work.